Move handle_image, get_news_info to fetch_data.py
