+++ zhihudaily/views/with_image.py
@@ -7,7 +7,7 @@ import datetime
 
 from flask import render_template, Blueprint
 
-from zhihudaily.utils import make_request, get_news_info, handle_image
+from zhihudaily.utils import make_request
 from zhihudaily.cache import cache
